update: restructure job api
